requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
ultralytics>=8.4.33
opencv-python-headless==4.12.0.88
pydantic-settings==2.11.0
